reorganize MatchFormation with subcomponents
